abstract away PKCE logic in BaseOAuth2PKCE for reuse
